Refactor cobbler sync
This will refactor cobbler sync by moving everything from litesync.py
into sync.py. Furthermore `make_tftpboot()` was made private and
renamed to `__create_tftpboot_dirs()` since it does only create the
required directories.

This is a preliminary work for #2601.
