+++ cobbler/actions/sync.py
@@ -25,6 +25,7 @@ import glob
 import logging
 import os
 import time
+from typing import Optional
 
 from cobbler.cexceptions import CX
 from cobbler import templar
@@ -55,6 +56,7 @@ class CobblerSync:
         self.distros = collection_mgr.distros()
         self.profiles = collection_mgr.profiles()
         self.systems = collection_mgr.systems()
+        self.images = collection_mgr.images()
         self.settings = collection_mgr.settings()
         self.repos = collection_mgr.repos()
         self.templar = templar.Templar(collection_mgr)
@@ -63,9 +65,6 @@ class CobblerSync:
         self.dhcp = dhcp
         self.tftpd = tftpd
         self.bootloc = self.settings.tftpboot_location
-        self.tftpgen.verbose = verbose
-        self.dns.verbose = verbose
-        self.dhcp.verbose = verbose
 
         self.pxelinux_dir = os.path.join(self.bootloc, "pxelinux.cfg")
         self.grub_dir = os.path.join(self.bootloc, "grub")
@@ -73,6 +72,9 @@ class CobblerSync:
         self.yaboot_bin_dir = os.path.join(self.bootloc, "ppc")
         self.yaboot_cfg_dir = os.path.join(self.bootloc, "etc")
         self.rendered_dir = os.path.join(self.settings.webdir, "rendered")
+        # FIXME: See https://github.com/cobbler/cobbler/issues/2453
+        # Move __create_tftpboot_dirs() outside of sync.py.
+        self.__create_tftpboot_dirs()
 
     def run(self):
         """
@@ -136,9 +138,9 @@ class CobblerSync:
         utils.run_triggers(self.api, None, "/var/lib/cobbler/triggers/sync/post/*")
         utils.run_triggers(self.api, None, "/var/lib/cobbler/triggers/change/*")
 
-    def make_tftpboot(self):
+    def __create_tftpboot_dirs(self):
         """
-        Make directories for tftpboot images
+        Create directories for tftpboot images
         """
         if not os.path.exists(self.pxelinux_dir):
             utils.mkdir(self.pxelinux_dir)
@@ -180,8 +182,7 @@ class CobblerSync:
                          "rendered"]:
                     # clean out directory contents
                     utils.rmtree_contents(path)
-        #
-        self.make_tftpboot()
+        self.__create_tftpboot_dirs()
         utils.rmtree_contents(self.pxelinux_dir)
         utils.rmtree_contents(self.grub_dir)
         utils.rmtree_contents(self.images_dir)
@@ -250,3 +251,160 @@ class CobblerSync:
         }
 
         self.templar.render(template_data, metadata, "/etc/rsyncd.conf")
+
+    def add_single_distro(self, name):
+        """
+        Sync adding a single distro.
+
+        :param name: The name of the distribution.
+        """
+        # get the distro record
+        distro = self.distros.find(name=name)
+        if distro is None:
+            return
+        # copy image files to images/$name in webdir & tftpboot:
+        self.tftpgen.copy_single_distro_files(distro, self.settings.webdir, True)
+        self.tftpd.add_single_distro(distro)
+
+        # create the symlink for this distro
+        src_dir = utils.find_distro_path(self.settings, distro)
+        dst_dir = os.path.join(self.settings.webdir, "links", name)
+        if os.path.exists(dst_dir):
+            self.logger.warning("skipping symlink, destination (%s) exists", dst_dir)
+        elif utils.path_tail(os.path.join(self.settings.webdir, "distro_mirror"), src_dir) == "":
+            self.logger.warning("skipping symlink, the source (%s) is not in %s",
+                                src_dir, os.path.join(self.settings.webdir, "distro_mirror"))
+        else:
+            try:
+                self.logger.info("trying symlink %s -> %s", src_dir, dst_dir)
+                os.symlink(src_dir, dst_dir)
+            except (IOError, OSError):
+                self.logger.error("symlink failed (%s -> %s)", src_dir, dst_dir)
+
+        # generate any templates listed in the distro
+        self.tftpgen.write_templates(distro, write_file=True)
+        # cascade sync
+        kids = distro.get_children()
+        for k in kids:
+            self.add_single_profile(k.name, rebuild_menu=False)
+        self.tftpgen.make_pxe_menu()
+
+    def add_single_image(self, name):
+        """
+        Sync adding a single image.
+
+        :param name: The name of the image.
+        """
+        image = self.images.find(name=name)
+        self.tftpgen.copy_single_image_files(image)
+        kids = image.get_children()
+        for k in kids:
+            self.add_single_system(k.name)
+        self.tftpgen.make_pxe_menu()
+
+    def remove_single_distro(self, name):
+        """
+        Sync removing a single distro.
+
+        :param name: The name of the distribution.
+        """
+        bootloc = self.settings.tftpboot_location
+        # delete contents of images/$name directory in webdir
+        utils.rmtree(os.path.join(self.settings.webdir, "images", name))
+        # delete contents of images/$name in tftpboot
+        utils.rmtree(os.path.join(bootloc, "images", name))
+        # delete potential symlink to tree in webdir/links
+        utils.rmfile(os.path.join(self.settings.webdir, "links", name))
+
+    def remove_single_image(self, name):
+        """
+        Sync removing a single image.
+
+        :param name: The name of the image.
+        """
+        bootloc = self.settings.tftpboot_location
+        utils.rmfile(os.path.join(bootloc, "images2", name))
+
+    def add_single_profile(self, name: str, rebuild_menu: bool = True) -> Optional[bool]:
+        """
+        Sync adding a single profile.
+
+        :param name: The name of the profile.
+        :param rebuild_menu: Whether to rebuild the grub/... menu or not.
+        :return: ``True`` if this succeeded.
+        """
+        # get the profile object:
+        profile = self.profiles.find(name=name)
+        if profile is None:
+            # Most likely a subprofile's kid has been removed already, though the object tree has not been reloaded and
+            # this is just noise.
+            return
+        # Rebuild the yum configuration files for any attached repos generate any templates listed in the distro.
+        self.tftpgen.write_templates(profile)
+        # Cascade sync
+        kids = profile.get_children()
+        for k in kids:
+            if k.COLLECTION_TYPE == "profile":
+                self.add_single_profile(k.name, rebuild_menu=False)
+            else:
+                self.add_single_system(k.name)
+        if rebuild_menu:
+            self.tftpgen.make_pxe_menu()
+        return True
+
+    def remove_single_profile(self, name: str, rebuild_menu: bool = True):
+        """
+        Sync removing a single profile.
+
+        :param name: The name of the profile.
+        :type name: str
+        :param rebuild_menu: Whether to rebuild the grub/... menu or not.
+        :type rebuild_menu: bool
+        """
+        # delete profiles/$name file in webdir
+        utils.rmfile(os.path.join(self.settings.webdir, "profiles", name))
+        # delete contents on autoinstalls/$name directory in webdir
+        utils.rmtree(os.path.join(self.settings.webdir, "autoinstalls", name))
+        if rebuild_menu:
+            self.tftpgen.make_pxe_menu()
+
+    def update_system_netboot_status(self, name: str):
+        """
+        Update the netboot status of a system.
+
+        :param name: The name of the system.
+        """
+        self.tftpd.update_netboot(name)
+
+    def add_single_system(self, name: str):
+        """
+        Sync adding a single system.
+
+        :param name: The name of the system.
+        """
+        # get the system object:
+        system = self.systems.find(name=name)
+        if system is None:
+            return
+        # rebuild system_list file in webdir
+        if self.settings.manage_dhcp:
+            self.dhcp.regen_ethers()
+        if self.settings.manage_dns:
+            self.dns.regen_hosts()
+        # write the PXE files for the system
+        self.tftpd.add_single_system(system)
+
+    def remove_single_system(self, name: str):
+        """
+        Sync removing a single system.
+
+        :param name: The name of the system.
+        """
+        bootloc = self.settings.tftpboot_location
+        # delete contents of autoinsts_sys/$name in webdir
+        system_record = self.systems.find(name=name)
+
+        for (name, interface) in list(system_record.interfaces.items()):
+            filename = system_record.get_config_filename(interface=name)
+            utils.rmfile(os.path.join(bootloc, "pxelinux.cfg", filename))
+            utils.rmfile(os.path.join(bootloc, "grub", filename.upper()))
