+++ cobbler/cobbler_collections/images.py
@@ -11,7 +11,6 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
 02110-1301  USA.
 """
 
-from cobbler.actions import litesync
 from cobbler.cobbler_collections import collection
 from cobbler.items import image as image
 from cobbler import utils
@@ -70,7 +69,7 @@ class Images(collection.Collection):
                 if with_triggers:
                     utils.run_triggers(self.collection_mgr.api, obj, "/var/lib/cobbler/triggers/delete/image/pre/*", [])
                 if with_sync:
-                    lite_sync = litesync.CobblerLiteSync(self.collection_mgr)
+                    lite_sync = self.collection_mgr.api.get_sync()
                     lite_sync.remove_single_image(name)
 
             self.lock.acquire()
