+++ cobbler/cobbler_collections/profiles.py
@@ -18,7 +18,6 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
 02110-1301  USA
 """
 
-from cobbler.actions import litesync
 from cobbler.cobbler_collections import collection
 from cobbler.items import profile as profile
 from cobbler import utils
@@ -86,7 +85,7 @@ class Profiles(collection.Collection):
                                        [])
                     utils.run_triggers(self.collection_mgr.api, obj, "/var/lib/cobbler/triggers/change/*", [])
                 if with_sync:
-                    lite_sync = litesync.CobblerLiteSync(self.collection_mgr)
+                    lite_sync = self.collection_mgr.api.get_sync()
                     lite_sync.remove_single_profile(name)
             return
