+++ cobbler/cobbler_collections/systems.py
@@ -18,7 +18,6 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
 02110-1301  USA
 """
 
-from cobbler.actions import litesync
 from cobbler.cobbler_collections import collection
 from cobbler.items import system as system
 from cobbler import utils
@@ -61,7 +60,7 @@ class Systems(collection.Collection):
                 if with_triggers:
                     utils.run_triggers(self.collection_mgr.api, obj, "/var/lib/cobbler/triggers/delete/system/pre/*", [])
                 if with_sync:
-                    lite_sync = litesync.CobblerLiteSync(self.collection_mgr)
+                    lite_sync = self.collection_mgr.api.get_sync()
                     lite_sync.remove_single_system(name)
             self.lock.acquire()
             try:
